requirements.txt
aiogram==3.4.1
aiofiles==23.2.1
aiosqlite==0.19.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.12.0
qrcode[pil]==7.4.2
pillow==10.2.0
geopy==2.4.1